ProyecAdmin/django_app/requirements.txt
Django==3.2.7
mysqlclient==2.1.0 
django-markdownx==4.0.0
Markdown==3.4.3